requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
coverage==7.9.2
cryptography==45.0.5
dotenv==0.9.9
ecdsa==0.19.1
Faker==24.2.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0